+++ R/cache.R
@@ -228,13 +228,13 @@ check_cache <- function() {
     is_too_old <- (Sys.time() - date_oldest_file) > "30 days"
   }
   if (is_too_big || is_too_old) {
-    cli::cli_alert_info(sprintf(
-      paste0(
-        "Cache dir: %s - size: %.0f MB - oldest file from: %s.\n",
-        "Clean up files older than 30 days with: `rcrisp::clear_cache('%s')` ",
-        "(or remove all cached files with: `rcrisp::clear_cache()`)."
-      ),
-      cache_dir, cache_size, as.Date(date_oldest_file), Sys.Date() - 30
+    cli::cli_alert_info(paste0(
+      "Cache dir: {cache_dir}",
+      " - size: {round(cache_size)} MB",
+      " - oldest file from: {as.Date(date_oldest_file)}.",
+      " Clean up files older than 30 days with:",
+      " `rcrisp::clear_cache('{Sys.Date() - 30}')`",
+      " (or remove all cached files with: `rcrisp::clear_cache()`)."
     ))
   }
 }